moved replace logic to shared
