+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/shared/resources/textResources/replace/replaceTextResourcesSagas.ts
@@ -4,6 +4,7 @@ import { IFormData } from "src/features/form/data/formDataReducer";
 import { IRuntimeState } from 'src/types';
 import { ITextResource } from "altinn-shared/types";
 import * as FormDataActionTypes from '../../../../features/form/data/formDataActionTypes';
+import { replaceTextResourceParams } from "altinn-shared/utils/language";
 export const FormDataSelector: (store: IRuntimeState) => IFormData = (store) => store.formData.formData;
 export const TextResourcesSelector: (store: IRuntimeState) => ITextResource[] = (store) => store.textResources.resources;
 
@@ -11,40 +12,10 @@ export const TextResourcesSelector: (store: IRuntimeState) => ITextResource[] = 
 export function* parseText(): SagaIterator {
   const formData: IFormData = yield select(FormDataSelector);
   const resource: ITextResource[] = yield select(TextResourcesSelector);
-  replaceTextResourceParams(resource, formData);
-
+  replaceTextResourceParams(resource, {'dataModel': formData});
 }
 
 export function* watchFetchFormDataFulfilled(): SagaIterator {
  yield takeLatest(FormDataActionTypes.FETCH_FORM_DATA_FULFILLED , parseText);
  yield takeLatest(FormDataActionTypes.UPDATE_FORM_DATA_FULFILLED, parseText);
 }
-
-function replaceTextResourceParams(textResources: ITextResource[], formData: IFormData): void {
-  textResources.forEach(resource => {
-    if(resource.variables != null){
-      var replaceValues: string[] = [];
-      resource.variables.forEach(variable => {
-        if(variable.dataSource.startsWith('dataModel')){
-          replaceValues.push(formData[variable.key] ? formData[variable.key] : variable.key);
-        }
-      });
-
-      var newValue: string = replaceParametersNullIndex(resource.unparsedValue, replaceValues);
-      if(resource.value != newValue){
-        resource.value = newValue;
-      }
-
-      resource.value =replaceParametersNullIndex(resource.unparsedValue, replaceValues);
-    }
-  });
-}
-
-  const replaceParametersNullIndex = (nameString: any, params: any[]) => {
-    let index = 0;
-    for (const param of params) {
-      nameString = nameString.replace(`{${index}}`, param);
-      index++;
-    }
-    return nameString;
-  };
